+++ wiki/Changelog.md
@@ -11,6 +11,9 @@ Terminal features
   * Fix OSC 52 response clipboard indication (#1327).
   * Fix emoji background in reverse mode.
 
+Character handling
+  * Workaround (for cygwin 3.6.4) for non-BMP characters in UTF-8 mode (#1325, #1326).
+
 Keyboard handling
   * Avoid sticky Ctrl state also after closing menu (~#1273, ~#1266).